requirements.txt
python-dotenv>=0.19.2
PyGithub>=1.55
pandas>=1.4.4
session_info>=1.0.0
regex>=2022.3.15